Move webpack configuration to a subfolder in the client directory
