+++ client/webpack/base.config.js
@@ -30,8 +30,8 @@ function entryPoints(paths) {
 
 
 module.exports = function exports() {
-  var CLIENT_DIR = path.resolve(__dirname, 'src');
-
+  var CLIENT_DIR = path.resolve(__dirname, '..', 'src');
+  
   return {
     entry: entryPoints('./wagtail/**/static_src/**/app/*.entry.js'),
     resolve: {
