+++ package.json
@@ -47,10 +47,10 @@
   },
   "scripts": {
     "postinstall": "cd ./client; npm install; cd ..",
-    "build": "gulp build; webpack --progress --colors --config ./client/webpack.prd.config.js",
-    "watch": "webpack --progress --colors --config ./client/webpack.dev.config.js & gulp watch",
+    "build": "gulp build; webpack --progress --colors --config ./client/webpack/prd.config.js",
+    "watch": "webpack --progress --colors --config ./client/webpack/dev.config.js & gulp watch",
     "start": "npm run watch",
-    "lint:js": "eslint --max-warnings 16 ./client/webpack.*.config.js ./client/src",
+    "lint:js": "eslint --max-warnings 16 ./client/webpack/*.config.js ./client/src",
     "lint": "npm run lint:js",
     "test": "npm run test:unit",
     "test:unit": "env NODE_PATH=$NODE_PATH:$PWD/client/src mocha --compilers js:babel-core/register client/tests/**/*.test.js",
